requirements.txt
streamlit>=1.35.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
sqlalchemy>=2.0.0
pandas>=2.2.0
python-dotenv>=1.0.0
plotly>=5.22.0
tabulate>=0.9.0
PyMySQL>=1.1.0
